Changed structure of tags & their types
- Tags is now a map where key = tag name & value = tag type name. I originally
considered to make tags their own class by having a set of tags which contain
their name, and a pointer to some tag type data stored in a separate set. This
would essentially, be the same result as a map, except to tag type data for a
particular tag would become O(n) instead of O(1), which is why I went the map
route.

- Tag types received a similar treatment as tags where tag types will now be
stored as a map where key = tag type name & value = tag type data (color).

- Tags also used to be in a read-only set so that it would only represent all
the tags in every song. If you were to remove a tag from a song and it was the
last tag to exist within a song, it would automatically be removed. While this
was initially was intended to be a feature, I now changed it so that tags can be
added and removed from the tags map directly. Instead, there will be a separate
feature to either 1) remove all unused tags, or 2) find all unused tags.

- Right now the tags, tag types, and collections they exist inside of are all
immutable, and therefore, thread-safe. Converting these collections to JSON
should be fairly straightforward as well, but has not yet been implemented.

- Tests still need to be written for most of this functionality.